tests/wpt/web-platform-tests/html/browsers/offline/application-cache-api/api_update_error.https.html

<!DOCTYPE HTML>
<html manifest="./non-existent-file.manifest">
  <head>
    <title>Offline Application Cache - API_update_error</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>

    <script>
      var cache = window.applicationCache;
      test(function() {
        assert_throws("INVALID_STATE_ERR", function(){cache.update();});
      }, "INVALID_STATE_ERR error test")
    </script>

  </body>
</html>
