LayoutTests/imported/w3c/web-platform-tests/html/browsers/offline/application-cache-api/api_update_error.https-expected.txt

CONSOLE MESSAGE: line 2: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: Application Cache manifest could not be fetched, because the manifest had a 404 response.

PASS INVALID_STATE_ERR error test 
